Norminette

+++ libft/src_ft_extras/ft_abs.c
@@ -6,13 +6,13 @@
 /*   By: okraus <okraus@student.42prague.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/19 15:42:56 by okraus            #+#    #+#             */
-/*   Updated: 2023/03/19 15:44:01 by okraus           ###   ########.fr       */
+/*   Updated: 2024/05/10 16:34:19 by okraus           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/libft.h"
 
-int	ft_abs(int n)
+long long	ft_abs(long long n)
 {
 	if (n < 0)
 		return (-n);
