+++ libft/src_ft_extras/ft_min.c
@@ -6,13 +6,13 @@
 /*   By: okraus <okraus@student.42prague.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/19 15:43:10 by okraus            #+#    #+#             */
-/*   Updated: 2023/03/19 15:43:35 by okraus           ###   ########.fr       */
+/*   Updated: 2024/05/10 16:29:03 by okraus           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/libft.h"
 
-int	ft_min(int a, int b)
+long long	ft_min(long long a, long long b)
 {
 	if (a < b)
 		return (a);
